modify `arkdep-build` to facilitate running in container
Part of this is about running in a container, and another part of it is being consistent about checking space in the places we are actually going to build and export. I ran into this when I was trying to build on a specific block device because I was working on a host with limited space in the root partition.